Switching from LibAuraInfo to internal
LibAuraInfo used e.g., 6sec for Kidney Shot, but since its duration varies that makes no sense in this context, in PlateBuffs it's OK if it's wrong, but here it cannot ever be wrong as that would cause the DR tracker to go haywire thinking there's been an early dynamic DR reset.

To fill in the list later.